Added a SWFUpload plugin which uses SWFObject to embed the Flash Object using the swfobject.EmbedSWF method.
The SWFUpload implementation uses innerHtml to modifiy the HTML code and insert the object element, but it fails in IE9 and Cloudfront. In this case, IE9 do not send the HTTP GET request until 120 seconds after inserting the Flash object element in the page.

Using this embedding mechanism, that weird problem disappears. Moreover, SWFObject is highly used for this purpose.